chore: modernize toolchain and add resume variants
- migrate from yarn to npm; drop .yarn/, yarn.lock
- upgrade React 18→19, Tailwind 3→4, Vite 4→8, TS 5→6
- replace SCSS/postcss/tailwind configs with Tailwind v4 vite plugin
- rework prerender: ssr.tsx + prerender.js → single prerender.tsx
- add src/variants/ with generic.json; src/resume.json becomes a symlink
- typography/layout refinements for tighter single-page A4 output

+++ src/formatters/cv-date-format.ts
@@ -1,5 +1,5 @@
 const cvDateFormat = (date: Date) => {
-  var options = { year: 'numeric', month: 'long' } as const;
+  var options = { year: 'numeric', month: 'short' } as const;
   return new Intl.DateTimeFormat('en', options).format(date);
 };
